Refactor sys fs functions into separate module
